Update almost all vim plugins
For some reason, YouCompleteMe does not allow me to commit the changes
in that repo, but is prevented by a "dirty" repo status?
Edit: YouCompleteMe seems to have update fine. For some reason it just
got stuck with some untracked files in the directory. After deleting
those, I could verify that the changes had been brought into this
branch.